refactor: audit scripts for CI/CD and address PR feedback
- Refactor `npm run audit` in `package.json` to correctly propagate exit status without masking errors.
- Move `ai-debugger.mjs` to `scripts/` and add environment variable support for target URLs.
- Update `audit_capture.py`, `ux-capture.js`, and `capture-screenshots.mjs` to use `BASE_URL` or equivalent environment variables.
- Remove redundant `accent-brand` token from `detect-antipatterns.mjs` and `Text` layout component.
- Simplify search highlighting in `GlobalSearch.tsx` by removing the `useSearchHighlight` hook and using inline logic to reduce indirection.
- Add audit artifacts (`antipattern-report.txt`, `TODO_ANTIPATTERNS.md`) to `.gitignore`.
