+++ scripts/ai-debugger.mjs
